+++ src/Chess/Domain/GameRepositoryITF.java
@@ -3,5 +3,5 @@ package Chess.Domain;
 public interface GameRepositoryITF
 {
     void save(Game g);
-    Game load(int id);
+    Game load(long id);
 }
